Separate code generating a response from `ProjectsController` to `ProjectsRespondable` concern
