+++ app/controllers/v3/auth/projects_controller.rb
@@ -1,48 +1,15 @@
 module V3
   module Auth
     class ProjectsController < ApplicationController
+      include ProjectsAccessible
+      include ProjectsRespondable
+
       before_action :validate_token_header!
 
       def index
-        credentials = UnifiedCredentials.new(Utils::Tokenator.from_token(request.headers[token_header_key]))
-        credentials_projects = credentials.groups.map { |group| group[:id] }
-        cloud_projects = Clouds::CloudProxy.new.projects
-
-        respond_with response_projects(credentials_projects & cloud_projects)
-      end
-
-      private
-
-      def response_projects(projects)
-        {
-          links: {
-            self: '',
-            previous: nil,
-            next: nil
-          },
-          projects: projects_hash(projects)
-        }
-      end
-
-      def projects_hash(projects)
-        projects.map { |project| project_hash project }
-      end
-
-      def project_hash(project)
-        {
-          is_domain: false,
-          description: '',
-          links: project_links_hash,
-          enabled: true,
-          id: project,
-          parent_id: '',
-          domain_id: '',
-          name: project
-        }
-      end
-
-      def project_links_hash
-        { self: '' }
+        @credentials = UnifiedCredentials.new(Utils::Tokenator.from_token(request.headers[x_auth_token_header_key]))
+        @cloud = Clouds::CloudProxy.new
+        respond_with projects_response
       end
     end
   end
